elasticsearch slicer can recuse on different key sets, removed more duplicated code
