+++ lib/readers/elasticsearch_date_range/slicer.js
@@ -1,18 +1,18 @@
 var Promise = require('bluebird');
 var moment = require('moment');
+var parser = require('datemath-parser');
+var _ = require('lodash');
 var processInterval = require('./../../utils/elastic_utils').processInterval;
 var buildQuery = require('./../../utils/elastic_utils').buildQuery;
 var buildRangeQuery = require('./../../utils/elastic_utils').buildRangeQuery;
 var dateOptions = require('./../../utils/elastic_utils').dateOptions;
 var dateFormatMS = require('./../../utils/elastic_utils').dateFormat;
 var dateFormatS = require('./../../utils/elastic_utils').dateFormatSeconds;
-var parser = require('datemath-parser');
-var _ = require('lodash');
+var getKeyArray = require('../../utils/id_utils').getKeyArray;
+var swapLastTwo = require('../../utils/id_utils').swapLastTwo;
+var transformIds = require('../../utils/id_utils').transformIds;
 var event = require('../../utils/events');
 
-var base64url = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w',
-    'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
-    'Y', 'Z', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '\-', '_'];
 
 function newSlicer(context, job, retryData, client) {
     var opConfig = job.readerConfig;
@@ -221,23 +221,18 @@ function newSlicer(context, job, retryData, client) {
             })
     }
 
-    function swapLastTwo(str) {
-        var last = str.charAt(str.length - 1);
-        var secondToLast = str.charAt(str.length - 2);
-
-        return str.slice(0, -2) + last + secondToLast
-    }
-
     function makeKeyList(opConfig, data) {
         var multiplier = opConfig.subslice_key_multiplier;
         var results = [];
+        var keyArray = getKeyArray(opConfig);
+        var transformedKeyArray = transformIds(opConfig, keyArray);
 
         function recurse(arr) {
             return Promise.all(arr.map(function(key) {
                     return getCount(opConfig, data, key)
                         .then(function(count) {
                             if (count >= opConfig.size * multiplier) {
-                                return recurse(base64url.map(function(str) {
+                                return recurse(keyArray.map(function(str) {
                                     return swapLastTwo(key + str)
                                 }))
                             }
@@ -256,11 +251,10 @@ function newSlicer(context, job, retryData, client) {
             );
         }
 
-        return recurse(base64url.map(function(key) {
-            return opConfig.type + '#' + key + '*'
-        })).then(function() {
-            return results;
-        });
+        return recurse(transformedKeyArray)
+            .then(function() {
+                return results;
+            });
     }
 
     function nextChunk(opConfig, client, jobConfig, dates, retryData) {
