+++ lib/readers/id_reader.js
@@ -2,9 +2,9 @@
 
 var Promise = require('bluebird');
 var _ = require('lodash');
-var base64url = require('../utils/id_utils').base64url;
-var hexadecimal = require('../utils/id_utils').hexadecimal;
 var swapLastTwo = require('../utils/id_utils').swapLastTwo;
+var getKeyArray = require('../utils/id_utils').getKeyArray;
+var transformIds = require('../utils/id_utils').transformIds;
 var getClient = require('../utils/config').getClient;
 var emitter = require('../utils/events');
 var Queue = require('../utils/queue');
@@ -128,20 +128,8 @@ function newSlicer(context, job, retryData) {
         })
     }
 
-    function transformIds(opConfig, arr) {
-        return arr.map(function(key) {
-            return opConfig.type + '#' + key + '*'
-        })
-    }
-
-    function getKeyArray(opConfig) {
-        if (opConfig.key_type === 'base64url') {
-            return  base64url;
-        }
-        else {
-            return hexadecimal;
-        }
-    }
+    
+    
 
     slicers.push(getID(getKeyArray(opConfig)));
