+++ lib/utils/id_utils.js
@@ -4,7 +4,7 @@ var base64url = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm'
     'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
     'Y', 'Z', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '\-', '_'];
 
-var hexadecimal = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9','A', 'B', 'C', 'D', 'E', 'F'];
+var hexadecimal = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'];
 
 
 function swapLastTwo(str) {
@@ -14,43 +14,25 @@ function swapLastTwo(str) {
     return str.slice(0, -2) + last + secondToLast
 }
 
-function makeKeyList(opConfig, data) {
-    var multiplier = opConfig.subslice_key_multiplier;
-    var results = [];
-
-    function recurse(arr) {
-        return Promise.all(arr.map(function(key) {
-                return getCount(opConfig, data, key)
-                    .then(function(count) {
-                        if (count >= opConfig.size * multiplier) {
-                            return recurse(base64url.map(function(str) {
-                                return swapLastTwo(key + str)
-                            }))
-                        }
-                        else {
-                            if (count !== 0) {
-                                results.push({
-                                    start: data.start.format(dateFormat),
-                                    end: data.end.format(dateFormat),
-                                    count: count,
-                                    key: key
-                                });
-                            }
-                        }
-                    })
-            })
-        );
+function getKeyArray(opConfig) {
+    if (opConfig.key_type === 'base64url') {
+        return base64url;
     }
+    else {
+        return hexadecimal;
+    }
+}
 
-    return recurse(base64url.map(function(key) {
+function transformIds(opConfig, arr) {
+    return arr.map(function(key) {
         return opConfig.type + '#' + key + '*'
-    })).then(function() {
-        return results;
-    });
+    })
 }
 
 module.exports = {
     base64url: base64url,
     hexadecimal: hexadecimal,
-    swapLastTwo: swapLastTwo
+    getKeyArray: getKeyArray,
+    swapLastTwo: swapLastTwo,
+    transformIds: transformIds
 };
